CCD-961 Merge Controller Advices in Def Store (#992)
* initial commit

* remove code smells
