+++ rest-api/src/test/java/uk/gov/hmcts/ccd/definition/store/rest/endpoint/CaseDefinitionControllerTest.java
@@ -16,7 +16,6 @@ import uk.gov.hmcts.ccd.definition.store.domain.service.casetype.CaseTypeVersion
 import uk.gov.hmcts.ccd.definition.store.repository.model.CaseType;
 import uk.gov.hmcts.ccd.definition.store.repository.model.Jurisdiction;
 import uk.gov.hmcts.ccd.definition.store.repository.model.Version;
-import uk.gov.hmcts.ccd.definition.store.rest.endpoint.exceptions.RestEndPointExceptionHandler;
 
 import java.util.List;
 import java.util.Optional;
@@ -49,7 +48,7 @@ public class CaseDefinitionControllerTest {
     public void createSubject() {
         subject = new CaseDefinitionController(caseTypeService, jurisdictionService, caseRoleService);
         mockMvc = MockMvcBuilders.standaloneSetup(subject)
-            .setControllerAdvice(new RestEndPointExceptionHandler())
+            .setControllerAdvice(new ControllerExceptionHandler())
             .build();
     }
